requirements.txt
# updated 11-25-2019 bml
confluent_kafka==1.3.0
nipyapi>=0.14.0
numpy>=1.17.4
pandas>=0.25.3
python-dateutil>=2.8.1
py4j>=0.10.8.1
pyspark>=2.4.4
PyYAML>=5.2
redis>=3.3.11
service>=0.6.0
splunk-sdk>=1.6.11
ua-parser>=0.8.0
user-agents>=2.0
# yappi>=1.2.1
findspark>=1.3.0
